move MozBrowserEvent to script_traits (+ documentation)
