+++ components/script/dom/htmliframeelement.rs
@@ -26,12 +26,11 @@ use dom::virtualmethods::VirtualMethods;
 use dom::window::Window;
 use js::jsapi::{JSAutoCompartment, JSAutoRequest, RootedValue, JSContext, MutableHandleValue};
 use js::jsval::{UndefinedValue, NullValue};
-use msg::constellation_msg::{ConstellationChan, MozBrowserEvent};
+use msg::constellation_msg::{ConstellationChan};
 use msg::constellation_msg::{NavigationDirection, PipelineId, SubpageId};
 use page::IterablePage;
-use script_traits::IFrameLoadInfo;
 use script_traits::IFrameSandboxState::{IFrameSandboxed, IFrameUnsandboxed};
-use script_traits::ScriptMsg as ConstellationMsg;
+use script_traits::{IFrameLoadInfo, MozBrowserEvent, ScriptMsg as ConstellationMsg};
 use std::ascii::AsciiExt;
 use std::cell::Cell;
 use string_cache::Atom;
