+++ components/script/dom/htmllinkelement.rs
@@ -24,10 +24,10 @@ use encoding::all::UTF_8;
 use ipc_channel::ipc;
 use ipc_channel::router::ROUTER;
 use layout_interface::{LayoutChan, Msg};
-use msg::constellation_msg::{ConstellationChan, MozBrowserEvent};
+use msg::constellation_msg::ConstellationChan;
 use net_traits::{AsyncResponseListener, AsyncResponseTarget, Metadata};
 use network_listener::{NetworkListener, PreInvoke};
-use script_traits::ScriptMsg as ConstellationMsg;
+use script_traits::{MozBrowserEvent, ScriptMsg as ConstellationMsg};
 use std::ascii::AsciiExt;
 use std::borrow::ToOwned;
 use std::cell::Cell;
